speed up

+++ app/Http/Controllers/ArticleController.php
@@ -12,6 +12,10 @@ class ArticleController extends Controller
 
     public function index(): View
     {
+        if (! app()->environment('local')) {
+            abort(403);
+        }
+
         return view('articles.index', [
             'articles' => Article::all(),
         ]);
@@ -19,11 +23,19 @@ class ArticleController extends Controller
 
     public function create(): View
     {
+        if (! app()->environment('local')) {
+            abort(403);
+        }
+
         return view('articles.create');
     }
 
     public function store(Request $request): RedirectResponse
     {
+        if (! app()->environment('local')) {
+            abort(403);
+        }
+
         $article = Article::create($request->all());
 
         return redirect()->route("articles.public_show", $article->slug);
@@ -41,6 +53,10 @@ class ArticleController extends Controller
 
     public function show(Article $article): View
     {
+        if (! app()->environment('local')) {
+            abort(403);
+        }
+
         return view('articles.show', [
             'article' => $article,
         ]);
@@ -48,6 +64,10 @@ class ArticleController extends Controller
 
     public function edit(Article $article): View
     {
+        if (! app()->environment('local')) {
+            abort(403);
+        }
+
         return view('articles.edit', [
             'article' => $article,
         ]);
@@ -55,6 +75,10 @@ class ArticleController extends Controller
 
     public function update(Request $request, Article $article): RedirectResponse
     {
+        if (! app()->environment('local')) {
+            abort(403);
+        }
+
         $article->update($request->all());
 
         return redirect()->route("articles.public_show", $article->slug);
@@ -62,6 +86,10 @@ class ArticleController extends Controller
 
     public function destroy(Article $article): RedirectResponse
     {
+        if (! app()->environment('local')) {
+            abort(403);
+        }
+
         $article->delete();
 
         return redirect()->route("articles.index");
